Extract pickleddb only tests
